long, long long, really looong! :-)

+++ query.h
@@ -57,5 +57,5 @@ int query_purge_deleted(MYSQL *mysql, long inode);
 
 int query_fsck(MYSQL *mysql);
 
-unsigned long query_total_inodes(MYSQL *mysql);
-unsigned long query_total_blocks(MYSQL *mysql);
+unsigned long long query_total_inodes(MYSQL *mysql);
+unsigned long long query_total_blocks(MYSQL *mysql);
